# TPBookstore-api/config/swagger.yaml
openapi: 3.0.1
info:
  title: npsstore API
  description: This is a Backend API of NBSStore
  version: 1.0.0
  termsOfService: http://quocbaoit.com/terms/
  contact:
    email: NBSstoreProject@quocbaoit.com
...
paths:
  /api/v1/product:
    post: #done
      description: Create a new product
      summary: Create a new product
      tags:
        - Product
      requestBody:
        description: Payload to create a new product
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductForm'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get: #done
      description: Get a part of products to show on shop client (non-user, user, admin)
      summary: Get a part of products to show on shop client (non-user, user, admin)
      tags:
        - Product
      parameters:
        - in: query
          name: pageSize
          required: false
          schema:
            type: number
          description: Number of products in 1 page
        - in: query
          name: pageNumber
          required: false
          schema:
            type: number
          description: Page number
        - in: query
          name: keyword
          required: false
          schema:
            type: string
          description: Keyword of product name to find in DB
        - in: query
          name: dateOrder
          required: false 
          schema:
            type: string
          description: Sort products by date (newest/latest)
        - in: query
          name: priceOrder
          required: false
          schema: 
            type: string
          description: Sort products by price (asc/desc)
        - in: query
          name: category
          required: false
          schema:
            type: string
          description: Product category name
        - in: query
          name: status
          required: false
          schema:
            type: string
          description: Sort products by status (disabled/notDisabled/all)        
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: number
                    description: current product page number
                  pages:
                    type: number
                    description: total product pages
                  products:
                    type: array
                    description: total product in 1 page
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: number
                    description: total products                   
        # '204':
        #   description: No products found for this keyword
        #   content:
        #     application/json:
        #       schema:
        #         $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  # /api/v1/product/all: #done
  #   get:
  #     summary: Return all products of store
  #     description: This api requires authorization with token JWT
  #     tags:
  #       - Product
  #     security:
  #       - bearerAuth: []
  #     responses:
  #       '200':
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               type: array
  #               items:
  #                 $ref: '#/components/schemas/Product'
  #       '401':
  #         description: Unauthorized
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/Error'
  /api/v1/product/{id}: #done
    get: 
      summary: Get matched product with given id (product detail)
      description: Get matched product with given id (product detail)
      tags:
        - Product
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put: #done
      summary: Update existed product
      description: Update existed product
      tags:
        - Product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: string
      requestBody:
        description: Payload to update product
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductForm'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete existed product
      description: Delete existed product
      tags:
        - Product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: error message
        stack:
          type: string
          description: error message description
    Product:
      type: object
      properties:
        _id:
          type: string
          description: Product ID
        name:
          type: string
          description: Product name
        image:
          type: string
          description: Product image's url
        description:
          type: string
          description: Product description
        reviews:
          type: array
          description: User reviews
          items:
            $ref: '#/components/schemas/Review'
        rating:
          type: number
          description: User ratings (5-stars rating)
        numReviews:
          type: number
          description: Number of reviews
        price:
          type: number
          description: Product price
        countInStock:
          type: number
          description: Total amount of a product available
        category:
          type: string
          description: Product category id
        totalSales:
          type: number
          description: Product total sales
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        __v:
          type: number
          description: Document version key
        isDisabled:
          type: boolean
          description: is Disabled or not
...
    ProductForm:
      type: object
      properties:
        name:
          type: string
          description: Product ID
        price:
          type: number
          description: Product price
        description:
          type: string
          description: Product description
        image:
          type: string
          description: Product image's url
        countInStock:
          type: number
          description: Total amount of a product available
        category:
          type: string
          description: Product category ID
    Timestamp:
      type: object
      properties:
        createAt:
          type: string
          description: The date and time when this record is created
          format: date-time
        updatedAt:
          type: string
          description: The date and time when this record is updated
          format: date-time
    Review:
      type: object
      properties:
        rating:
          type: integer
          description: User ratings (5-stars rating)
          minimum: 1
          maximum: 5
        reviewContent:
          type: string
          description: Review content          
        user:
          type: string
          description: User ID
        timestamp:
          $ref: '#/components/schemas/Timestamp'
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      description: "<i>JWT Authorization header using the Bearer scheme.</i> <b>Example</b>: <i><u>Bearer 12345abcdef</u></i>"
      in: header          # can be "header", "query" or "cookie"
      name: Authorization # name of the header, query parameter or cookie
# 2) Apply the security globally to all operations
security:
  - bearerAuth: [] # use the same name as above